# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Live Music Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/preferences:
    get:
      operationId: getUserPreferences
      tags: [users]
      summary: Get user location preferences
      responses:
        "200":
          description: User preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
    put:
      operationId: updateUserPreferences
      tags: [users]
      summary: Update user location preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserPreferencesBody"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
...
  /shows:
    get:
      operationId: listShows
      tags: [shows]
      summary: List shows filtered by user preferences or location
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: zipCode
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: venueId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of shows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShowWithDetails"
  /shows/{id}:
    get:
      operationId: getShow
      tags: [shows]
      summary: Get a show by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Show details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShowWithDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shows/{id}/attend:
    post:
      operationId: markAttending
      tags: [attendance]
      summary: Mark the current user as attending a show
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceBody"
      responses:
        "200":
          description: Attendance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
    delete:
      operationId: removeAttendance
      tags: [attendance]
      summary: Remove attendance for the current user from a show
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /shows/{id}/attendees:
    get:
      operationId: getShowAttendees
      tags: [attendance]
      summary: Get list of users attending a show
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of attendees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendeeInfo"
...
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by username
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSearchResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - isAdmin
        - createdAt
    UserPreferences:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        cities:
          type: array
          items:
            type: string
        zipCodes:
          type: array
          items:
            type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - cities
        - zipCodes
        - updatedAt
    UpdateUserPreferencesBody:
      type: object
      properties:
        cities:
          type: array
          items:
            type: string
        zipCodes:
          type: array
          items:
            type: string
      required:
        - cities
        - zipCodes
    Venue:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        state:
          type: ["string", "null"]
        zipCode:
          type: ["string", "null"]
        websiteUrl:
          type: string
        scrapeUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        lastScrapedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - city
        - websiteUrl
        - isActive
        - createdAt
...
    ShowWithDetails:
      type: object
      properties:
        id:
          type: integer
        venueId:
          type: integer
        title:
          type: string
        artist:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        showDate:
          type: string
        doorsTime:
          type: ["string", "null"]
        showTime:
          type: ["string", "null"]
        ticketUrl:
          type: ["string", "null"]
        ticketPrice:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        venue:
          $ref: "#/components/schemas/Venue"
        attendeeCount:
          type: integer
        currentUserAttending:
          type: boolean
        currentUserBoughtTickets:
          type: boolean
        friendsAttending:
          type: array
          items:
            $ref: "#/components/schemas/AttendeeInfo"
      required:
        - id
        - venueId
        - title
        - showDate
        - createdAt
        - venue
        - attendeeCount
        - currentUserAttending
        - currentUserBoughtTickets
        - friendsAttending
    Attendance:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        showId:
          type: integer
        boughtTickets:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - showId
        - boughtTickets
        - createdAt
    AttendanceBody:
      type: object
      properties:
        boughtTickets:
          type: boolean
      required:
        - boughtTickets
    AttendeeInfo:
      type: object
      properties:
        userId:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        boughtTickets:
          type: boolean
      required:
        - userId
        - username
        - boughtTickets
...
    UserSearchResult:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        isFriend:
          type: boolean
        hasPendingRequest:
          type: boolean
      required:
        - id
        - username
        - isFriend
        - hasPendingRequest
...