# lib/api-zod/src/generated/types/attendance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Live Music Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Attendance {
  id: number;
  userId: number;
  showId: number;
  boughtTickets: boolean;
  createdAt: string;
}