# lib/api-zod/src/generated/types/attendeeInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Live Music Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface AttendeeInfo {
  userId: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  boughtTickets: boolean;
}