# lib/api-zod/src/generated/types/showWithDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Live Music Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendeeInfo } from "./attendeeInfo";
import type { Venue } from "./venue";

export interface ShowWithDetails {
  id: number;
  venueId: number;
  title: string;
  /** @nullable */
  artist?: string | null;
  /** @nullable */
  description?: string | null;
  showDate: string;
  /** @nullable */
  doorsTime?: string | null;
  /** @nullable */
  showTime?: string | null;
  /** @nullable */
  ticketUrl?: string | null;
  /** @nullable */
  ticketPrice?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  sourceUrl?: string | null;
  createdAt: string;
  venue: Venue;
  attendeeCount: number;
  currentUserAttending: boolean;
  currentUserBoughtTickets: boolean;
  friendsAttending: AttendeeInfo[];
}