# lib/api-zod/src/generated/types/venue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Live Music Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Venue {
  id: number;
  name: string;
  city: string;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zipCode?: string | null;
  websiteUrl: string;
  /** @nullable */
  scrapeUrl?: string | null;
  isActive: boolean;
  /** @nullable */
  lastScrapedAt?: string | null;
  createdAt: string;
}